% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mp_class.R
\name{hcr}
\alias{hcr}
\alias{hcr,simpleMP-method}
\alias{hcr<-}
\alias{hcr<-,simpleMP,character-method}
\alias{hcr_params}
\alias{hcr_params,simpleMP-method}
\alias{hcr_params<-}
\alias{hcr_params<-,simpleMP,numeric-method}
\alias{get_hcr_shape}
\alias{get_hcr_shape,simpleMP-method}
\title{Accessor methods}
\usage{
hcr(object, ...)

\S4method{hcr}{simpleMP}(object)

hcr(object, ...) <- value

\S4method{hcr}{simpleMP,character}(object) <- value

hcr_params(object, ...)

\S4method{hcr_params}{simpleMP}(object)

hcr_params(object, ...) <- value

\S4method{hcr_params}{simpleMP,numeric}(object) <- value

get_hcr_shape(object, ...)

\S4method{get_hcr_shape}{simpleMP}(object, hcr_ip)
}
\description{
Accessor methods
}
