% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bet}
\alias{bet}
\title{simpleBiol object for bigeye projections}
\format{
An object of class simpleBiol with slots:
\describe{
\item{n}{Numbers in the population. \code{FLQuant}.}
\item{n0}{Numbers in the population with no fishing. \code{FLQuant}.}
\item{m}{Mortality rate of the population. \code{FLQuant}.}
\item{wt}{Mean weight of an individual. \code{FLQuant}.}
\item{mat}{Proportion of individuals mature. \code{FLQuant}.}
\item{rec_dist}{Distribution of total annual recruitment between seasons and areas. \code{FLQuant}.}
\item{srr_params}{SRR parameters. \code{FLPar}.}
\item{movement}{Movement rates between model areas. \code{array}.}
\item{name}{Name of the object. \code{character}.}
\item{desc}{Brief description of the object. \code{character}.}
}
}
\description{
simpleBiol object for bigeye projections
}
