% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bet_fish}
\alias{bet_fish}
\title{simpleFisheries object for bigeye projections}
\format{
An object of class simpleFisheries with slots:
\describe{
\item{catch_n}{Catch numbers. \code{FLQuant}.}
\item{catch_wt}{Catch weights at age. \code{FLQuant}.}
\item{sel}{Selectivity. \code{FLQuant}.}
\item{catch_q}{Catchability paramsters. \code{FLQuant}.}
\item{effort}{Fishing effort. \code{FLQuant}.}
\item{name}{Name of the object. \code{character}.}
\item{desc}{Brief description of the object. \code{character}.}
}
}
\description{
simpleFisheries object for bigeye projections
}
