% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_effort}
\alias{find_effort}
\title{Finds the effort that achieves the supplied catch target.}
\usage{
find_effort(
  n_pre_move,
  m,
  waa,
  movement,
  selq,
  effort_mult_initial,
  target,
  target_type,
  fishery_area,
  max_effort,
  max_solver_iters
)
}
\arguments{
\item{n_pre_move}{a 2D array of numbers-at-age by area}

\item{m}{natural a 2D array of natural mortality-at-age by area}

\item{waa}{a 2D array of weight-at-age by fishery}

\item{movement}{a 3D array of movement-at-age}

\item{selq}{a 2D array of selectivity-at-age by fishery}

\item{effort_mult_initial}{numeric initial guess of effort multiplier for solver}

\item{target}{numeric vector of targets by fishery}

\item{target_type}{integer vector of target types by fishery (0 = catch, 1 = effort)}

\item{fishery_area}{integer vector indicating which area each fishery operates in}

\item{max_effort}{numeric vector of maxmimum effort by fishery}

\item{max_solver_iters}{integer maximum number of solver iterations}

\item{n0_pre_move}{a 2D array of unfished numbers-at-age by area}
}
\value{
list of solve_effort().
}
\description{
find_effort() is a wrapper for solve_effort() that is exposed to R.
}
