% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recruitment.R
\name{get_annual_recruitment}
\alias{get_annual_recruitment}
\alias{get_annual_recruitment-method}
\title{Annual recruitment}
\usage{
get_annual_recruitment(object, ...)
}
\arguments{
\item{year}{The year recruitment is to be calculated for.}

\item{srr_devs}{An FLQuant of total annual recruitment deviates.}

\item{zero_effort}{If TRUE, use n0 slot to recruitment with no fishing.}
}
\description{
Calculates recruitment in all seasons and areas of a year.
Based on the MFCL method of getting total SSB in year-1, calculating total recruitment in year,
applying a deviate, then splitting the resulting total annual recruitment across the seasons and areas in year.
}
\examples{
\dontrun{
data(bet_projection_low_catch_bits_simple)
annual_rec <- get_annual_recruitment(bet, year = ycount, srr_devs = srr_devs)
}
}
