% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_error}
\alias{get_error}
\title{Defines f(x) for effort minimisation.}
\arguments{
\item{nfisheries_}{an integer number of fisheries}

\item{n_after_move}{a 2D array of numbers-at-age by area}

\item{m}{natural a 2D array of natural mortality-at-age by area}

\item{waa}{a 2D array of weight-at-age by fishery}

\item{selq}{a 2D array of selectivity-at-age by fishery}

\item{target}{numeric vector of targets by fishery}

\item{target_type}{integer vector of target types by fishery (0 = catch, 1 = effort)}

\item{fishery_map}{integer vector indicating which area each fishery operates in}
}
\value{
f(x) evaluated at x.
}
\description{
For catch fisheries:
\deqn{g(x) = \sum [log(C_{target}) - log(C(x))]^2}
For effort fisheries:
\deqn{h(x) = \sum [log(E_{target}) - log(x)]^2}
Error function is then \eqn{f(x) = \sqrt{g(x) + h(x)}}.
}
