% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recruitment.R
\name{get_recruitment_orig}
\alias{get_recruitment_orig}
\title{Get recruitment}
\usage{
get_recruitment_orig(pop_n, waa, mat, srr_params, srr_devs, rec_dist)
}
\arguments{
\item{pop_n}{Numbers at age in each season of year y}

\item{waa}{Weights at age}

\item{mat}{Maturity at age}

\item{srr_params}{SRR parameters}

\item{srr_devs}{Deviates in year y}

\item{rec_dist}{Proportion of total annual recruitment in each season and region}
}
\description{
Get recruitment in all seasons and regions of year y+1, given n etc in year y
}
