% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survivors_and_catch.R
\name{get_survivors_and_catch_orig}
\alias{get_survivors_and_catch_orig}
\title{Get survivors and catch numbers at age in a time step}
\usage{
get_survivors_and_catch_orig(
  effort,
  pop_n,
  m,
  movement,
  sel,
  catch_q,
  fishery_area
)
}
\arguments{
\item{effort}{Vector of effort}

\item{pop_n}{N at age}

\item{m}{Natural mortality at age}

\item{movement}{Movement rates between model regions}

\item{sel}{Selectivity at age}

\item{catch_q}{Catchability}

\item{fishery_area}{Vector of which model region each fishery operates in}
}
\value{
A list of survivors and catch numbers at age
}
\description{
Follows population and fishing dynamics from Multfan-CL.
Movement, followed by death (natural mortality and fishing).
}
