% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mp_class.R
\name{hcr_threshold}
\alias{hcr_threshold}
\alias{hcr_asymptotic}
\alias{hcr_constant}
\alias{hcr_hillary_step}
\alias{hcr_constrained}
\alias{hcr_asymptotic_constrained}
\alias{hcr_threshold_constrained}
\alias{hcr_hillary_step_constrained}
\alias{hcr_asymptotic_hillary_step}
\alias{hcr_asymptotic_hillary_step_constrained}
\title{hcr_threshold}
\usage{
hcr_threshold(hcr_ip, params)

hcr_asymptotic(hcr_ip, params)

hcr_constant(hcr_ip, params)

hcr_hillary_step(hcr_ip, params)

hcr_constrained(hcr_ip, params, reference_out, hcr_func)

hcr_asymptotic_constrained(hcr_ip, params, reference_out)

hcr_threshold_constrained(hcr_ip, params, reference_out)

hcr_hillary_step_constrained(hcr_ip, params, reference_out)

hcr_asymptotic_hillary_step(hcr_ip, params)

hcr_asymptotic_hillary_step_constrained(hcr_ip, params, reference_out)
}
\arguments{
\item{hcr_ip}{Numeric vector of HCR inputs}

\item{params}{Numeric vector of HCR parameters.}

\item{reference_out}{The reference level that the constraint is applied to.}

\item{hcr_func}{Character string of the function name}
}
\value{
Numberic vector of the HCR output

Numberic vector of the HCR output

Numberic vector of the HCR output

Numberic vector of the HCR output
}
\description{
Typical HCR function with the parameters hcr_ip_min, hcr_ip_max, out_min, out_max passed as a vector.
The default HCR for the MFCLMSEControl class.

Instead of a threshold shape this HCR has a lovely curvy shape based on an asymptotic function.

An HCR function that just returns the same output irrespective of the SBSBF0 input

Hillary step is basically two thresholds on top of each other, side by side
Parameters are 'hcr_ip_min', 'hcr_ip_max', 'hcr_ip_step_min', 'hcr_ip_step_max','out_min', 'out_max', 'step_height'.

Takes any HCR function as a character string and applies constraint if necessary

Instead of a threshold shape this HCR has a lovely curvy shape based on an asymptotic function.
Also has a maximum constraint change up or down from previous value

Like the typical HCR function with the parameters but with a constraint on how much the output is allowed to change from a reference output.
Parameters are 'hcr_ip_min', 'hcr_ip_max', 'out_min', 'out_max', 'max_change_up', 'max_change_down'.
There is an additional argument 'reference_out' that is the reference output, i.e. the value to which to the new output is compared.

Like the typical HCR function with the parameters but with a constraint on how much the output is allowed to change from a reference output.
Parameters are 'hcr_ip_min', 'hcr_ip_max', 'out_min', 'out_max', 'max_change_up', 'max_change_down'.
There is an additional argument 'reference_out' that is the reference output, i.e. the value to which to the new output is compared.

Hillary step is basically two thresholds on top of each other, side by side
Parameters are 'hcr_ip_min', 'hcr_ip_max', 'hcr_ip_step_min', 'hcr_ip_step_max','out_min', 'out_max', 'step_height'.
Adding 'curve' to allow for asymptotic out_left

Like the typical HCR function with the parameters but with a constraint on how much the output is allowed to change from a reference output.
Parameters are 'hcr_ip_min', 'hcr_ip_max', 'out_min', 'out_max', 'max_change_up', 'max_change_down'.
There is an additional argument 'reference_out' that is the reference output, i.e. the value to which to the new output is compared.
}
