% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project}
\alias{project}
\title{Projects fishery with catch or effort targets.}
\usage{
project(
  n_pre_move,
  n0_pre_move,
  m,
  waa,
  movement,
  selq,
  target,
  target_type,
  fishery_area,
  max_effort,
  max_solver_iters
)
}
\arguments{
\item{n_pre_move}{a 2D array of numbers-at-age by area}

\item{n0_pre_move}{a 2D array of unfished numbers-at-age by area}

\item{m}{natural a 2D array of natural mortality-at-age by area}

\item{waa}{a 2D array of weight-at-age by fishery}

\item{movement}{a 3D array of movement-at-age}

\item{selq}{a 2D array of selectivity-at-age by fishery}

\item{target}{vector of targets by fishery}

\item{target_type}{vector of target types by fishery (0 = catch, 1 = effort)}

\item{fishery_area}{integer vector indicating which area each fishery operates in}

\item{max_effort}{numeric vector of maxmimum effort by fishery}

\item{max_solver_iters}{integer maximum number of solver iterations}
}
\value{
list of effort, solver info and updated population quantities
}
\description{
project() is mostly a wrapper for solve_effort() but also applies movement to the numbers-at-age arrays and calculates survivors and catch. Fisheries with catch targets require optimisation of effort to acheive target catch. solve_effort() is an internal function that can be accessed through the exposed function find_effort().
}
