% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_methods.R
\name{SB,FLQuant,numeric,numeric-method}
\alias{SB,FLQuant,numeric,numeric-method}
\alias{SB,FLQuant,missing,missing-method}
\alias{SB,FLQuant,numeric,missing-method}
\alias{SB,FLQuant,missing,numeric-method}
\alias{sb_methods}
\alias{SB}
\alias{SB,simpleBiol,missing,missing-method}
\alias{SB,simpleBiol,numeric,missing-method}
\alias{SB,simpleBiol,missing,numeric-method}
\alias{SBF0}
\alias{SBF0,simpleBiol,numeric,numeric-method}
\alias{SBF0,simpleBiol,missing,missing-method}
\alias{SBF0,simpleBiol,numeric,missing-method}
\alias{SBF0,simpleBiol,missing,numeric-method}
\alias{SBSBF0}
\alias{SBSBF0,simpleBiol,numeric,numeric,numeric,numeric-method}
\alias{SBSBF0,simpleBiol,missing,missing,missing,missing-method}
\alias{SBrecent}
\alias{SBrecent,simpleBiol-method}
\alias{SBlatest}
\alias{SBlatest,simpleBiol-method}
\alias{SBF0recent}
\alias{SBF0recent,simpleBiol-method}
\alias{SBSBF0recent}
\alias{SBSBF0recent,simpleBiol-method}
\alias{SBSBF0latest}
\alias{SBSBF0latest,simpleBiol-method}
\title{Methods for returns the spawning biomass (fished or unfished) and the depletion}
\usage{
\S4method{SB}{FLQuant,numeric,numeric}(object, mean_nyears, lag_nyears, combine_areas = TRUE, season_means = TRUE)

\S4method{SB}{FLQuant,missing,missing}(object, mean_nyears, lag_nyears, ...)

\S4method{SB}{FLQuant,numeric,missing}(object, mean_nyears, lag_nyears, ...)

\S4method{SB}{FLQuant,missing,numeric}(object, mean_nyears, lag_nyears, ...)

\S4method{SB}{simpleBiol,numeric,numeric}(object, mean_nyears, lag_nyears, combine_areas = TRUE, season_means = TRUE)

\S4method{SB}{simpleBiol,missing,missing}(object, mean_nyears, lag_nyears, ...)

\S4method{SB}{simpleBiol,numeric,missing}(object, mean_nyears, lag_nyears, ...)

\S4method{SB}{simpleBiol,missing,numeric}(object, mean_nyears, lag_nyears, ...)

SBF0(object, mean_nyears, lag_nyears, ...)

\S4method{SBF0}{simpleBiol,numeric,numeric}(
  object,
  mean_nyears,
  lag_nyears,
  combine_areas = TRUE,
  season_means = TRUE
)

\S4method{SBF0}{simpleBiol,missing,missing}(object, mean_nyears, lag_nyears, ...)

\S4method{SBF0}{simpleBiol,numeric,missing}(object, mean_nyears, lag_nyears, ...)

\S4method{SBF0}{simpleBiol,missing,numeric}(object, mean_nyears, lag_nyears, ...)

SBSBF0(
  object,
  sb_mean_nyears,
  sb_lag_nyears,
  sbf0_mean_nyears,
  sbf0_lag_nyears,
  ...
)

\S4method{SBSBF0}{simpleBiol,numeric,numeric,numeric,numeric}(
  object,
  sb_mean_nyears,
  sb_lag_nyears,
  sbf0_mean_nyears,
  sbf0_lag_nyears,
  ...
)

\S4method{SBSBF0}{simpleBiol,missing,missing,missing,missing}(
  object,
  sb_mean_nyears,
  sb_lag_nyears,
  sbf0_mean_nyears,
  sbf0_lag_nyears,
  ...
)

SBrecent(object, ...)

\S4method{SBrecent}{simpleBiol}(object, ...)

SBlatest(object, ...)

\S4method{SBlatest}{simpleBiol}(object, ...)

SBF0recent(object, ...)

\S4method{SBF0recent}{simpleBiol}(object, ...)

SBSBF0recent(object, ...)

\S4method{SBSBF0recent}{simpleBiol}(object, ...)

SBSBF0latest(object, ...)

\S4method{SBSBF0latest}{simpleBiol}(object, ...)
}
\arguments{
\item{object}{A single simpleBiol object}

\item{mean_nyears}{The number of years to calculate the rolling mean over. A value of 1 is the same as not calculating a mean. For the SB() method the default value is 1.}

\item{lag_nyears}{The number of years to lag the output. A value of 0 is the same as not lagging the output. For the SB() method the default value is 0.}

\item{combine_areas}{If TRUE the biomassses in different areas are summed. Default value is TRUE.}

\item{season_means}{If TRUE the seasonal biomassses are averaged in each year so only annual biomass is returned. Default value is TRUE.}

\item{...}{Other arguments.}

\item{sb_mean_nyears}{Equivalent to the argument \code{mean_nyears} for the SB part of the \code{SBSBF0()} method.}

\item{sb_lag_nyears}{Equivalent to the argument \code{lag_nyears} for the SB part of the \code{SBSBF0()} method.}

\item{sbf0_mean_nyears}{Equivalent to the argument \code{mean_nyears} for the SBF0 part of the \code{SBSBF0()} method.}

\item{sbf0_lag_nyears}{Equivalent to the argument \code{lag_nyears} for the SBF0 part of the \code{SBSBF0()} method.}
}
\value{
An FLQuant
}
\description{
Methods for returning the spawning biomass (\code{SB()}), unfished spawning biomass (\code{SBF0()}) and depletion (\code{SBSBF0()}) from a single simpleBiol object.
The biomass is averaged over seasons.
It is possible to calculate a rolling mean over a specified number of years
as well as lag the output.
These can specified separately for the SB and SBF0 parts of SBSBF0.
The number of years in the outputs will match that of the inputs.
The start of the returned FLQuant will be padded with NAs if necessary (for example, where the output has been lagged).
}
\details{
Shortcut methods are available (\code{SBlatest()}, \code{SBrecent()}, \code{SBF0recent()} and \code{SBSBF0recent()}). These set the \code{mean_nyears()} and \code{lag_nyears()} parameters to match those used in the assessment reports.
\code{SBlatest()} does not calculate a mean or lag the output.
\code{SBrecent()} uses a rolling mean of 4 years without a lag.
\code{SBF0recent()} uses a rolling mean of 10 years with a lag of 1 year.
\code{SBSBF0recent()} is a combination of \code{SBrecent()} and \code{SBF0recent()}.
\code{SBSBF0latest()} is a combination of \code{SBlatest()} and \code{SBF0recent()}.
}
