% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_classes.R, R/constructors.R,
%   R/accessors.R
\docType{class}
\name{simpleBiol}
\alias{simpleBiol}
\alias{simpleBiol-class}
\alias{n,simpleBiol-method}
\alias{m,simpleBiol-method}
\alias{wt,simpleBiol-method}
\alias{name,simpleBiol-method}
\alias{desc,simpleBiol-method}
\alias{n<-,simpleBiol,FLQuant-method}
\alias{m<-,simpleBiol,FLQuant-method}
\alias{wt<-,simpleBiol,FLQuant-method}
\alias{name<-,simpleBiol,character-method}
\alias{desc<-,simpleBiol,character-method}
\alias{simpleBiol-methods}
\alias{simpleBiol,FLQuant-method}
\alias{simpleBiol,missing-method}
\alias{simpleBiolcpp-class}
\alias{n<-,simpleBiol,numeric-method}
\alias{n0}
\alias{n0<-}
\alias{n0,simpleBiol-method}
\alias{n0<-,simpleBiol,numeric-method}
\alias{m<-,simpleBiol,numeric-method}
\alias{wt<-,simpleBiol,numeric-method}
\alias{mat,simpleBiol-method}
\alias{mat<-,simpleBiol,numeric-method}
\alias{rec_dist}
\alias{rec_dist<-}
\alias{rec_dist,simpleBiol-method}
\alias{movement,simpleBiol-method}
\alias{rec_dist<-,simpleBiol,numeric-method}
\alias{srr_params}
\alias{srr_params<-}
\alias{srr_params,simpleBiol-method}
\alias{srr_params<-,simpleBiol,numeric-method}
\alias{movement}
\alias{movement<-}
\alias{movement<-,simpleBiol,numeric-method}
\alias{iter,simpleBiol-method}
\title{Class simpleBiol}
\usage{
simpleBiol(object, ...)

\S4method{simpleBiol}{FLQuant}(object, ...)

\S4method{n}{simpleBiol}(object)

\S4method{n}{simpleBiol,numeric}(object) <- value

n0(object, ...)

n0(object, ...) <- value

\S4method{n0}{simpleBiol}(object)

\S4method{n0}{simpleBiol,numeric}(object) <- value

\S4method{m}{simpleBiol}(object)

\S4method{m}{simpleBiol,numeric}(object) <- value

\S4method{wt}{simpleBiol}(object)

\S4method{wt}{simpleBiol,numeric}(object) <- value

\S4method{mat}{simpleBiol}(object)

\S4method{mat}{simpleBiol,numeric}(object) <- value

rec_dist(object, ...)

rec_dist(object, ...) <- value

\S4method{rec_dist}{simpleBiol}(object)

\S4method{rec_dist}{simpleBiol,numeric}(object) <- value

srr_params(object, ...)

srr_params(object, ...) <- value

\S4method{srr_params}{simpleBiol}(object)

\S4method{srr_params}{simpleBiol,numeric}(object) <- value

movement(object, ...)

movement(object, ...) <- value

\S4method{movement}{simpleBiol}(object)

\S4method{movement}{simpleBiol,numeric}(object) <- value

\S4method{name}{simpleBiol}(object)

\S4method{desc}{simpleBiol}(object)

\S4method{iter}{simpleBiol}(obj, iter)
}
\description{
A simple class for modelling age structured populations.
}
\details{
The \code{simpleBiol} class is a representation of a biological fish population.
This includes information on abundances, natural mortality and maturity.
}
\section{Slots}{
 \describe{
\item{n}{Numbers in the population. \code{FLQuant}.}
\item{n0}{Numbers in the population with no fishing. \code{FLQuant}.}
\item{m}{Mortality rate of the population. \code{FLQuant}.}
\item{wt}{Mean weight of an individual. \code{FLQuant}.}
\item{mat}{Proportion of individuals mature. \code{FLQuant}.}
\item{rec_dist}{Distribution of total annual recruitment between seasons and areas. \code{FLQuant}.}
\item{srr_params}{SRR parameters. \code{FLPar}.}
\item{movement}{Movement rates between model areas. \code{array}.}
\item{name}{Name of the object. \code{character}.}
\item{desc}{Brief description of the object. \code{character}.}
}

\section{Validity}{
 \describe{
\item{Dimensions}{All FLQuant slots must have iters equal to 1 or 'n'.}
\item{Iters}{The dimname for iter\link{1} should be '1'.}
\item{Dimnames}{The name of the quant dimension must be the same for all FLQuant slots.}
}
}

\author{
Finlay Scott
}
\keyword{classes}
