% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_classes.R, R/constructors.R,
%   R/accessors.R
\docType{class}
\name{simpleFisheries}
\alias{simpleFisheries}
\alias{simpleFisheries-class}
\alias{simpleFisheries-methods}
\alias{simpleFisheries,FLQuant-method}
\alias{simpleFisheries,missing-method}
\alias{simpleFisheriescpp-class}
\alias{catch_n}
\alias{catch_n<-}
\alias{catch_n,simpleFisheries-method}
\alias{catch_n<-,simpleFisheries,numeric-method}
\alias{catch_wt}
\alias{catch_wt<-}
\alias{catch_wt,simpleFisheries-method}
\alias{catch_wt<-,simpleFisheries,numeric-method}
\alias{sel}
\alias{sel<-}
\alias{sel,simpleFisheries-method}
\alias{sel<-,simpleFisheries,numeric-method}
\alias{catch_q}
\alias{catch_q<-}
\alias{catch_q,simpleFisheries-method}
\alias{catch_q<-,simpleFisheries,numeric-method}
\alias{effort,simpleFisheries,ANY-method}
\alias{effort,simpleFisheries-method}
\alias{effort<-,simpleFisheries,numeric-method}
\alias{fishery_map}
\alias{fishery_map<-}
\alias{fishery_map,simpleFisheries-method}
\alias{name,simpleFisheries-method}
\alias{desc,simpleFisheries-method}
\alias{catch}
\alias{simpleBiol-methods}
\alias{catch,simpleFisheries-method}
\alias{iter,simpleFisheries-method}
\title{Class simpleFisheries}
\usage{
simpleFisheries(object, ...)

\S4method{simpleFisheries}{FLQuant}(object, ...)

catch_n(object, ...)

catch_n(object, ...) <- value

\S4method{catch_n}{simpleFisheries}(object)

\S4method{catch_n}{simpleFisheries,numeric}(object) <- value

catch_wt(object, ...)

catch_wt(object, ...) <- value

\S4method{catch_wt}{simpleFisheries}(object)

\S4method{catch_wt}{simpleFisheries,numeric}(object) <- value

sel(object, ...)

sel(object, ...) <- value

\S4method{sel}{simpleFisheries}(object)

\S4method{sel}{simpleFisheries,numeric}(object) <- value

catch_q(object, ...)

catch_q(object, ...) <- value

\S4method{catch_q}{simpleFisheries}(object)

\S4method{catch_q}{simpleFisheries,numeric}(object) <- value

\S4method{effort}{simpleFisheries,ANY}(object)

\S4method{effort}{simpleFisheries,numeric}(object) <- value

fishery_map(object, ...)

fishery_map(object, ...) <- value

\S4method{fishery_map}{simpleFisheries}(object)

\S4method{catch_q}{simpleFisheries,numeric}(object) <- value

\S4method{name}{simpleFisheries}(object)

\S4method{desc}{simpleFisheries}(object)

catch(object, ...)

\S4method{catch}{simpleFisheries}(object)

\S4method{iter}{simpleFisheries}(obj, iter)
}
\description{
A simple class for modelling fisheries.
}
\details{
The \code{simpleFisheries} class is a simple representation of multiple fisheries.
This includes information on abundances, natural mortality and maturity.
This class only works for fishing on a single stock.
Fisheries are stored in the unit dimension.
}
\section{Slots}{
 \describe{
\item{catch_n}{Catch numbers. \code{FLQuant}.}
\item{catch_wt}{Catch weights at age. \code{FLQuant}.}
\item{sel}{Selectivity. \code{FLQuant}.}
\item{catch_q}{Catchability paramsters. \code{FLQuant}.}
\item{effort}{Fishing effort. \code{FLQuant}.}
\item{name}{Name of the object. \code{character}.}
\item{desc}{Brief description of the object. \code{character}.}
}

\section{Validity}{
 \describe{
\item{Dimensions}{All FLQuant slots must have iters equal to 1 or 'n'.}
\item{Iters}{The dimname for iter\link{1} should be '1'.}
\item{Dimnames}{The name of the quant dimension must be the same for all FLQuant slots.}
}
}

\author{
Finlay Scott
}
\keyword{classes}
