% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mp_class.R
\name{simpleMP}
\alias{simpleMP}
\title{simpleMP}
\usage{
simpleMP(
  hcr = "hcr_threshold",
  hcr_params = c(hcr_ip_min = 0.2, hcr_ip_max = 0.5, out_min = 0.2, out_max = 1),
  ...
)
}
\description{
Basic constructor for MFCLMSEControl class
}
