% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_effort}
\alias{solve_effort}
\title{Finds the effort that achieves the supplied catch target.}
\usage{
solve_effort(
  n_pre_move,
  m,
  waa,
  movement,
  selq,
  target,
  target_type,
  fishery_map,
  max_effort,
  max_solver_iters
)
}
\arguments{
\item{n_pre_move}{a 2D array of numbers-at-age by area}

\item{m}{a 2D array of natural mortality-at-age by area}

\item{waa}{a 2D array of weight-at-age by fishery}

\item{movement}{a 3D array of movement-at-age}

\item{selq}{a 2D array of selectivity-at-age by fishery}

\item{target}{numeric vector of targets by fishery}

\item{target_type}{integer vector of target types by fishery (0 = catch, 1 = effort)}

\item{max_effort}{numeric vector of maxmimum effort by fishery}

\item{max_solver_iters}{integer maximum number of solver iterations}

\item{n0_pre_move}{a 2D array of unfished numbers-at-age by area}

\item{fishery_area}{integer vector indicating which area each fishery operates in}
}
\value{
list of effort, solver info and updated population quantities
}
\description{
Originally, solve_effort() was not exposed to R, but now it is exported to help with documentation.
Uses the LBFGSBSolver to perform minimisation. The minimised function, f(x), is of type EffortFun. See ?get_error.
}
